Add support for unknown batch sizes in group_norm.
Previously, the `group_norm` layer did not work when the batch size was not known (as is the case when using e.g. `tf.data.Dataset.batch`). 

The change directly addresses this case by replacing a call to `Tensor.shape` with a call to `array_ops.shape`, as well as replacing `None` with `-1` when calling `array_ops.reshape`.